beam_lib, compile: Replace use of deprecated crypto functions
Since both the STDLIB and compiler applications turn warnings
into errors, we must stop using the old deprecated crypto functions.

While we are at it, generalize the format of the key tuple returned
by beam_lib:make_crypto_key/2 to facilitate introducing new crypto
methods in the future. Change the format to:

  {Type,Key,IV,BlockSize}

where Type, Key, and IV are the first three arguments for either
crypto:block_encrypt4/ or crypto:block_decrypt/4, and BlockSize
is the block size for the crypto algorithm (it is needed to properly
pad the plaintext blocks before encryption).