dtls: rework lpain to cipher text encoding
For TLS the connection state of current and next are not sufficient.
When sending a flight, we might have laready processed a CCS half way
through the flight. During encoding, we now have to use connection
state information form a previous epoch.

This change add support for keeping the previouse epoch's connection
state and select it by Epoch Id.